pyproject.toml
[project]
name = "easy-mcp-proxy"
version = "0.1.0"
description = "Easy MCP Proxy - Filter, transform, and compose MCP tools"
readme = "README.md"
license = "AGPL-3.0-or-later"
authors = [
    {name = "Andrew Brookins", email = "a@andrewbrookins.com"}
]
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "wsproto>=1.3.2",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
mcp-proxy = "mcp_proxy.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_proxy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
addopts = "--cov=mcp_proxy --cov-report=term-missing --cov-fail-under=100"

[tool.coverage.run]
source = ["mcp_proxy"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
# Exclude partial branches that are hard to test
partial_branches = [
    "pragma: no branch",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

